% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{best.seed}
\alias{best.seed}
\title{Best seed function}
\usage{
best.seed(plsko.tuning, seed.measure = "TPP", target.fdr = 0.05)
}
\arguments{
\item{plsko.tuning}{The tuning result from plsko_tuning function.}

\item{seed.measure}{The measure to be used to select the best seed. Default is "TPP".}

\item{target.fdr}{The target FDR to be used as a dash line in the plot. Default is 0.05.}
}
\description{
Return the plsko.seed from the optimal configuration based on the either highest TPP when FDP is under the target, or the lowest FDP, or the lowest squared sum of FDP and (1-TPP)
}
\keyword{internal}
