% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{DGPy.AR1}
\alias{DGPy.AR1}
\title{Generate artificial response variable y}
\usage{
DGPy.AR1(X, s, A, c = 0, y.dis = "Normal")
}
\arguments{
\item{X}{The dataset to be used for tuning. A numeric matrix or data frame where rows represent observations and columns represent variables.}

\item{s}{The number of assumed important variables to be used to generate artifical .}

\item{A}{The range of the coefficients for the true support.}

\item{c}{The noise level. Default is 0.}

\item{y.dis}{The distribution of the response variable. Default is "Normal".}
}
\value{
A list containing the true support and the generated response variable y.
}
\description{
Generate artificial response variable y based on the true support and coefficients.
}
\keyword{internal}
