% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{fdr}
\alias{fdr}
\title{Calculate False discovery proportion}
\usage{
fdr(S, beta.true)
}
\arguments{
\item{S}{The selected variables.}

\item{beta.true}{The true support.}
}
\value{
The false discovery proportion (FDP).
}
\description{
Calculate the false discovery proportion (FDP) based on the selected variables and the true support.
}
\keyword{internal}
