% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_filter.R
\name{integer0_test}
\alias{integer0_test}
\title{Integer Zero Test}
\usage{
integer0_test(which)
}
\arguments{
\item{which}{A vector of indices.}
}
\value{
Either the input indices or `NULL` if the input is empty.
}
\description{
Helper function to handle empty selections.
}
\keyword{internal}
