% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_filter.R
\name{ko.filter}
\alias{ko.filter}
\title{Knockoff filter with generated knockoff variable set}
\usage{
ko.filter(X, Xk, y, q = 0.05, method = "lasso.lcd", offset = 0, ...)
}
\arguments{
\item{X}{A \eqn{n \times p} numeric matrix of predictors.}

\item{Xk}{A \eqn{n \times p} numeric matrix of knockoff predictors.}

\item{y}{A numeric or factor type repose vector of length \eqn{n}.}

\item{q}{A numeric value of the target false discovery rate (FDR) level.}

\item{method}{A string specifying the method to compute test statistics. Options include
`"lasso.lcd"`, `"lasso.max.lambda"`, `"lasso.logistic"`, `"RF"`. Default is `"lasso.lcd"`.}

\item{offset}{A numeric or string value to adjust the knockoff threshold. Default is `0` (control modifed FDR). Other options include `1` (yields a slightly more conservative procedure ("knockoffs+") that controls the FDR according to the usual definition) and `"both"` (returns results from both "knockoffs" and "knockoffs+").}

\item{...}{Additional arguments passed to the underlying methods. See functions `stat.lasso_coefdiff`, `stat.lasso_lambdadiff`, `stat.lasso_coefdiff_bin`, `stat.random_forest` in `knockoff` package for more details.}
}
\value{
An object of class "knockoff.result". This object is a list
 containing at least the following components:
\describe{
  \item{`call`}{The matched call.}
  \item{`X`}{The original matrix of predictors.}
  \item{`Xk`}{The knockoff matrix of predictors.}
  \item{`y`}{The response vector.}
  \item{`statistic`}{The computed test statistics `W` of length \eqn{p}.}
  \item{`threshold`}{The knockoff threshold used to determine discoveries.}
  \item{`selected`}{The named vector of selected variables.}
}
}
\description{
Knockoff filter with generated knockoff variable set
}
\examples{
# Example usage of ko.filter
result <- ko.filter(X, Xk, y, q = 0.1)

}
\references{
Candes et al., Panning for Gold: Model-free Knockoffs for High-dimensional Controlled Variable Selection,
  arXiv:1610.02351 (2016).
  \href{https://web.stanford.edu/group/candes/knockoffs/index.html}{https://web.stanford.edu/group/candes/knockoffs/index.html}

  Barber and Candes,
  Controlling the false discovery rate via knockoffs.
  Ann. Statist. 43 (2015), no. 5, 2055--2085.
}
