% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_filter.R
\name{ko.withW}
\alias{ko.withW}
\title{Knockoff Filtering with Precomputed Test Statistics}
\usage{
ko.withW(W, q = 0.05, offset = 0)
}
\arguments{
\item{W}{A numeric vector of test statistics.}

\item{q}{A numeric value specifying the false discovery rate (FDR) level. Default is `0.05`.}

\item{offset}{A numeric or string value to adjust the knockoff threshold. Default is `0` (control modifed FDR). Other options include `1` (yields a slightly more conservative procedure ("knockoffs+") that controls the FDR according to the usual definition) and `"both"` (returns results from both "knockoffs" and "knockoffs+").}
}
\value{
An object of class "knockoff.result". This object is a list
 containing at least the following components:
\describe{
  \item{`call`}{The matched call.}
  \item{`X`}{The original matrix of predictors.}
  \item{`Xk`}{The knockoff matrix of predictors.}
  \item{`y`}{The response vector.}
  \item{`statistic`}{The computed test statistics `W` of length \eqn{p}.}
  \item{`threshold`}{The knockoff threshold used to determine discoveries.}
  \item{`selected`}{The named vector of selected variables.}
}
}
\description{
Computes the knockoff threshold and selects variables based on the computed test statistics.
}
