% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSKO.R
\name{linear.regression.generator}
\alias{linear.regression.generator}
\title{Calculate \eqn{\hat{X}} by fitting OLS regression on its neighbours}
\usage{
linear.regression.generator(Y, X)
}
\description{
Calculate \eqn{\hat{X}} by fitting OLS regression on its neighbours
}
\keyword{internal}
