% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{plot.plsko.tuning}
\alias{plot.plsko.tuning}
\title{Plot the tuning result}
\usage{
\method{plot}{plsko.tuning}(plsko.tuning, target.fdr = 0.05)
}
\arguments{
\item{plsko.tuning}{The tuning result from plsko_tuning function.}

\item{target.fdr}{The target FDR to be used as a dash line in the plot. Default is 0.05.}
}
\description{
Plot the tuning result as a boxplot of FDP and TPP, with the optimal configuration highlighted with a 'star' on the correspondence coordinate in the plot.
}
