% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSKO.R
\name{pls.recovery.generator}
\alias{pls.recovery.generator}
\title{Calculate \eqn{\hat{X}} by fitting PLS regression on its neighbours}
\usage{
pls.recovery.generator(Y, X, ncomp, keepX = rep(ncol(X)))
}
\description{
Calculate \eqn{\hat{X}} by fitting PLS regression on its neighbours
}
\keyword{internal}
