% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsAKO.R
\name{plsAKO}
\alias{plsAKO}
\title{Integration Multiple PLSKO with Aggregate Knockoffs (AKO) Procedure}
\usage{
plsAKO(
  X,
  y,
  n_ko = 25,
  q = 0.05,
  offset = 1,
  w.method = "lasso.lcd",
  gamma = 0.3,
  seed = 1,
  parallel = T,
  cores = NULL,
  ...
)
}
\arguments{
\item{X}{A numeric matrix or dataframe. The original predictor data matrix with \eqn{n} observations as rows and \eqn{p} variables as columns.}

\item{y}{A numeric vector of responses.}

\item{n_ko}{An integer specifying the number of knockoff copies to generate by PLSKO. Default is 25.}

\item{q}{A numeric value specifying the target false discovery rate (FDR). Default is 0.05.}

\item{offset}{An integer (0 or 1) specifying the offset in the empirical p-value calculation. Default is 1.}

\item{w.method}{A character string specifying the method to compute feature importance statistics. Default is \code{"lasso.lcd"}. Other options include \code{"lasso.logistic"} for binary response variable, \code{"lasso.max.lambda"} for the maximum lambda value for the first entry on the path, and \code{"RF"} for random forest.}

\item{gamma}{A numeric value between 0 and 1 for the quantile aggregation parameter. Default is 0.3.}

\item{seed}{An integer to set the random seed for reproducibility. Default is 1.}

\item{parallel}{Logical value indicating whether to run the process in parallel. Default is TRUE.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Default is NULL, which uses all available cores except one.}

\item{...}{Additional arguments passed to the \link[=plsko]{plsko} knockoff-generating function.}
}
\value{
A list of class 'AKO.result', containing:
\describe{
  \item{`s`}{A list of results from individual knockoff filters.}
  \item{`ako.s`}{The named vector of selected variables after aggregating knockoff results.}
  \item{`threshold`}{The FDR control threshold from AKO.}
}
}
\description{
This function implements the Aggregate Knockoffs (AKO) procedure from Nguyen et al. (2020)  with knockoff variables generated by PLSKO, to address the randomness from single-run knockoffs and improve the stablity. It generates multiple knockoff copies from PLSKO, computes feature importance statistics, aggregates p-values across knockoff realisations to control the false discovery rate. Compared to single-run knockoffs, AKO improved the average power with FDR control. The function is adapted from Tian et al. (2022).
}
\examples{
\dontrun{
set.seed(123)
X <- matrix(rnorm(100 * 10), 100, 10)
y <- rnorm(100)
result <- plsAKO(X, y)
}

}
\references{
Nguyen T-B, Chevalier J-A, Thirion B et al. Aggregation of Multiple Knockoffs. Proceedings of the 37th International Conference on Machine Learning. PMLR, 2020, 7283–93.

Tian P, Hu Y, Liu Z et al. Grace-AKO: a novel and stable knockoff filter for variable selection incorporating gene network structures. BMC Bioinformatics 2022;23:478.
}
