% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_batch.R
\name{plsko_batch}
\alias{plsko_batch}
\title{Generate multiple PLSKO knockoff variables}
\usage{
plsko_batch(
  X,
  n_ko = 10,
  nb.list = NULL,
  threshold.abs = NULL,
  threshold.q = NULL,
  ncomp = NULL,
  sparsity = 1,
  rmax = 5,
  seed = Sys.time(),
  simpls = F
)
}
\arguments{
\item{X}{A numeric matrix or data frame where rows represent observations and columns represent variables.}

\item{n_ko}{An integer specifying the number of knockoff variables to generate. Default is 10.}
}
\value{
A list of length n_ko, each element is a matrix of n*p of knockoff variables.
}
\description{
This function generates multiple knockoff variables for a given design matrix using the PLS regression method with mutliple times residuals only. Please note this function is fast but does not generate rigidly valid knockoff variables under the definition.
}
\keyword{internal}
