% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{plsko_semi_sim}
\alias{plsko_semi_sim}
\title{Semi-simulation for PLSKO}
\usage{
plsko_semi_sim(
  X,
  p_s,
  q,
  n_ko,
  plsko.ncomp,
  plsko.threshold.abs,
  plsko.sparsity,
  simpls,
  seed = Sys.time(),
  parallel = TRUE,
  ncores = NULL
)
}
\arguments{
\item{X}{The dataset to be used for tuning. A numeric matrix or data frame where rows represent observations and columns represent variables.}

\item{p_s}{The number of assumed important variables to be used to generate artifical . Default is 10 \% of the number of variables.}

\item{q}{The target FDR to be used as a dash line in the plot. Default is 0.05.}

\item{n_ko}{The number of repetitions for the semi-simulation. Default is 10.}

\item{plsko.ncomp}{The number of components to be used in PLSKO.}

\item{plsko.threshold.abs}{The threshold for the absolute value of the correlation between the original and knockoff variables.}

\item{plsko.sparsity}{The sparsity level in the PLS regression.}

\item{simpls}{Logical. If TRUE, use the SIMPLS algorithm to fit the PLS model. Default is FALSE.}

\item{seed}{An integer seed for reproducibility. Default is 1.}

\item{parallel}{Logical. If TRUE, run the semi-simulation in parallel. Default is TRUE.}

\item{ncores}{An integer specifying the number of cores to use for parallel computation. Default is NULL.}
}
\value{
A list containing the result as a dataframe including the configuration, performance metrics (including FDP and TPP), run times and average performance metrics
}
\description{
Return FDP and TPP of semi-simulation for n_ko repetitions at this configuration
}
