% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{pow}
\alias{pow}
\title{Calculate True positive proportion}
\usage{
pow(S, beta.true)
}
\arguments{
\item{S}{The selected variables.}

\item{beta.true}{The true support.}
}
\value{
The true positive proportion (TPP).
}
\description{
Calculate the true positive proportion (TPP) based on the selected variables and the true support.
}
\keyword{internal}
