% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsAKO.R
\name{print.AKO.result}
\alias{print.AKO.result}
\title{Print AKO Result}
\usage{
\method{print}{AKO.result}(x, ...)
}
\arguments{
\item{x}{An object of class `'AKO.result'`, which contains the results from running the AKO procedure.}

\item{...}{Additional arguments (currently unused).}
}
\value{
None. The function prints the call and the set of selected variables.
}
\description{
Prints the results of the Aggregated Knockoff (AKO) filter, including the call and selected variables.
}
\examples{
# Example usage of print.AKO.result
print(result)

}
\seealso{
\code{\link{plsAKO}}, \code{\link{AKO_withKO}}, \code{\link{AKO_withW}}
}
