% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_filter.R
\name{print.knockoff.result}
\alias{print.knockoff.result}
\title{Print results for the knockoff filter}
\usage{
\method{print}{knockoff.result}(x, plus = F, ...)
}
\arguments{
\item{x}{the output of a call to ko.filter}

\item{...}{unused}
}
\description{
Prints the list of variables selected by the knockoff filter and the corresponding function call.
}
