% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsAKO.R
\name{quantile_aggregation}
\alias{quantile_aggregation}
\title{Quantile Aggregation}
\usage{
quantile_aggregation(pvals, gamma = 0.3)
}
\arguments{
\item{pvals}{A numeric vector of p-values.}

\item{gamma}{A numeric value for the quantile to use in aggregation. Default is `0.3`.}
}
\value{
A numeric value representing the aggregated p-value.
}
\description{
Aggregates multiple p-values using quantile aggregation based on Meinshausen et al. (2009). Function Adapted from P Tian et al. (2022).
}
\references{
Meinshausen N, Meier L, Bühlmann P. p-Values for High-Dimensional Regression. Journal of the American Statistical Association 2009;104:1671–81.

Tian P, Hu Y, Liu Z et al. Grace-AKO: a novel and stable knockoff filter for variable selection incorporating gene network structures. BMC Bioinformatics 2022;23:478.
}
\keyword{internal}
