% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_criterion.R
\name{r_criterion}
\alias{r_criterion}
\title{Calculate the Empirical Number of Components for PLS Regression}
\usage{
r_criterion(X, rmax = 10)
}
\arguments{
\item{X}{A numeric matrix or data frame where rows represent observations and columns represent variables.}

\item{rmax}{An integer specifying the maximum number of factors to consider. Default is 10.}
}
\value{
An integer representing the optimal number of components based on the \eqn{PC_p1} criterion.
}
\description{
This function calculates the empirical number of components used for PLS regression
by applying the \eqn{PC_p1} criterion from Bai and Ng (2002). It determines the optimal number of factors based on the minimum value of the criterion. The method and function is adapted from Y Fan et al. (2019).
}
\examples{
set.seed(123)
X <- matrix(rnorm(100*10), 100, 10)
optimal_r <- r_criterion(X)

}
\references{
Fan Y, Lv J, Sharifvaghefi M et al. IPAD: Stable Interpretable Forecasting with Knockoffs Inference. Journal of the American Statistical Association 2020;115:1822–34.

Bai J, Ng S. Determining the Number of Factors in Approximate Factor Models. Econometrica 2002;70:191–221.
}
