% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSKO.R
\name{simpls.recovery.generator}
\alias{simpls.recovery.generator}
\title{Calculate \eqn{\hat{X}} by fitting PLS regression on its neighbours using simpls algorithm}
\usage{
simpls.recovery.generator(Y, X, ncomp)
}
\description{
Calculate \eqn{\hat{X}} by fitting PLS regression on its neighbours using simpls algorithm
}
\keyword{internal}
