% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsAKO.R
\name{AKO_withKO}
\alias{AKO_withKO}
\title{Aggregated Knockoff (AKO) with Generated Knockoff Sets}
\usage{
AKO_withKO(
  X,
  y,
  Xko.list,
  q = 0.05,
  w.method = "lasso.lcd",
  offset = 1,
  gamma = 0.3,
  seed = 1
)
}
\arguments{
\item{X}{A \eqn{n \times p} numeric matrix or data frame of predictors.}

\item{y}{A numeric or factor response vector of length \eqn{n}.}

\item{Xko.list}{A list of knockoff variable matrices (\eqn{n \times p}) corresponding to the original matrix `X`.}

\item{q}{A numeric value of the target false discovery rate (FDR) level. Default is `0.05`.}

\item{w.method}{A string specifying the method to compute test statistics. Options are `"lasso.lcd"`, `"lasso.logistic"`, `"lasso.max.lambda"`, `"RF"`. Default is `"lasso.lcd"`.}

\item{offset}{A numeric value to specify the adjustment in empirical p-value calculation. Options are `0` or `1`. Default is `1`.}

\item{gamma}{A numeric value for quantile aggregation in the multiple knockoff p-value aggregation. Default is `0.3`.}

\item{seed}{A numeric value for the random seed. Default is `1`.}
}
\value{
A list of class 'AKO.result', containing:
\describe{
  \item{`s`}{A list of results from individual knockoff filters.}
  \item{`ako.s`}{The named vector of selected variables after aggregating knockoff results.}
  \item{`threshold`}{The FDR control threshold.}
}
}
\description{
Performs Aggregated Knockoff analysis with multiple knockoff sets to improve the stability of results.
}
\examples{
# Example usage of AKO_withKO
result <- AKO_withKO(X, y, Xko.list, q = 0.1)

}
\references{
Nguyen T-B, Chevalier J-A, Thirion B et al. Aggregation of Multiple Knockoffs. Proceedings of the 37th International Conference on Machine Learning. PMLR, 2020, 7283–93.

Tian P, Hu Y, Liu Z et al. Grace-AKO: a novel and stable knockoff filter for variable selection incorporating gene network structures. BMC Bioinformatics 2022;23:478.
}
