% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsAKO.R
\name{AKO_withW}
\alias{AKO_withW}
\title{Aggregated Knockoff with Precomputed Test Statistics \eqn{W} From User-Defined Methods}
\usage{
AKO_withW(W, q = 0.05, offset = 0, gamma = 0.3)
}
\arguments{
\item{W}{A list of test statistics from multiple knockoff filters.}

\item{q}{A numeric value of the target false discovery rate (FDR) level. Default is \eqn{0.05}.}

\item{offset}{An integer (0 or 1) specifying the offset in the empirical p-value calculation. Default is \eqn{0} (liberally control modified FDR with higher power). Other options include \eqn{1}, similar to "knockoffs+", yielding a slightly more conservative procedure that controls the FDR according to the usual definition.}

\item{gamma}{A numeric value for quantile aggregation in the multiple knockoff p-value aggregation. Default is \eqn{0.3}.}
}
\value{
A list containing:
\describe{
  \item{`s`}{A list of results from individual knockoff filters.}
  \item{`ako.s`}{The set of selected variables after aggregating knockoff results.}
  \item{`threshold`}{The FDR control threshold.}
}
}
\description{
Performs Aggregated Knockoff analysis using precomputed test statistics `W`.
}
\examples{
# Example usage of AKO_withW
result <- AKO_withW(W, q = 0.05)

}
