% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSKO.R
\name{plsko}
\alias{plsko}
\title{Generate a knockoff variable set with PLSKO using PLS regression}
\usage{
plsko(
  X,
  nb.list = NULL,
  threshold.abs = NULL,
  threshold.q = NULL,
  ncomp = NULL,
  sparsity = 1
)
}
\arguments{
\item{X}{A numeric matrix or data frame. The original design data matrix with \eqn{n} observations as rows and \eqn{p} variables as columns.}

\item{nb.list}{Optional. A list of length \eqn{p} or adjacency matrix of \eqn{p \times p} that defines the neighbourship of variables. A list of length \eqn{p} should include the neighbours' index of each variable from \eqn{X_1} to \eqn{X_p} in order; The \eqn{i^{th}} element in the list includes the indices of the neighbour variables of \eqn{X_i}, or \code{NULL} when no neighbours. A adjacency matrix should be symmetric with only binary element and  where \eqn{M_{ij} = 1} when \eqn{X_i} and \eqn{X_j} is defined as neighbours; otherwise \eqn{M_{ij} = 0} when not neighbour or on diagonal (i.e. \eqn{i = j}). If not provided or NULL, the neighborhoods are determined based on correlations.}

\item{threshold.abs}{Optional. A value between \eqn{0} and \eqn{1}. A numeric value specifying an absolute correlation threshold to define neighborhoods. If not provided, the function uses \code{threshold.q}.}

\item{threshold.q}{Optional. A numeric value between 0 and 1 indicating the quantile of the correlation values to use as a threshold.}

\item{ncomp}{Optional. An integer specifying the number of components to use in the PLS regression. Default is NULL, the \code{ncomp} is determined empirically by \eqn{PC_p1} criterion.}

\item{sparsity}{Optional. A numeric value between 0 and 1 specifying the sparsity level in the PLS regression. Default is 1 (no sparsity).}
}
\value{
A matrix of generated knockoff variables of \eqn{n \times p}
}
\description{
Generate a knockoff variable set with PLSKO using PLS regression
}
