% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsko_tuning.R
\name{plsko_tuning}
\alias{plsko_tuning}
\title{PLSKO tuning function}
\usage{
plsko_tuning(
  X,
  p_s = round(0.1 * ncol(X)),
  n_ko = 10,
  q = 0.05,
  parallel = TRUE,
  ncore = NULL,
  ncomp = seq(3, 9, 2),
  threshold.q = NULL,
  threshold.abs = NULL,
  sparsity = 1,
  simpls = F,
  seed = 1,
  fdp.measure = "median",
  early.stop = F
)
}
\arguments{
\item{X}{The dataset to be used for tuning. A numeric matrix or data frame where rows represent observations and columns represent variables.}

\item{p_s}{The number of assumed important variables to be used to generate artificial y. Default is 10 \% of the number of variables.}

\item{n_ko}{The number of repetitions for the semi-simulation. Default is 10.}

\item{q}{The target FDR level. Default is 0.05.}

\item{parallel}{Logical. If TRUE, run the semi-simulation in parallel. Default is TRUE.}

\item{ncore}{An integer specifying the number of cores to use for parallel computation. Default is NULL.}

\item{ncomp}{A integer or a vector of integers to be tested, specifying the number of components to be used in PLSKO. Default is seq(3, 9, 2).}

\item{threshold.q}{A numeric value or a vector of numeric values (between 0 - 1) to be tested, specifying a quantile threshold to define neighborhoods. Default is NULL. This parameter will be used if threshold.abs is not provided.}

\item{threshold.abs}{A numeric value or a vector of numeric values (between 0 - 1) to be tested, specifying an absolute correlation threshold to define neighborhoods. Default is NULL.}

\item{sparsity}{A numeric value or a vector of numeric values (between 0 (excluded) - 1) to be tested, specifying the sparsity level in the PLS regression. Default is 1.}

\item{simpls}{Logical. If TRUE, use the SIMPLS algorithm to fit the PLS model. Default is FALSE.}

\item{seed}{An integer seed for reproducibility. Default is 1.}

\item{fdp.measure}{A character specifying the measure to be used to select the optimal configuration. Once the mean or median FDP is below the target FDR, the configuration will be return as optimal. Default is "median". Options are "mean", "median", "either" and "both".}

\item{early.stop}{Logical. If TRUE, the tuning will stop once the mean or median FDP is below the target FDR. Default is FALSE.}
}
\value{
An object of class "plsko.tuning". This object is a list containing at least the following components:
\describe{
 \item{`full`}{The full result of the tuning.}
 \item{`mean`}{The average result of the tuning.}
 \item{`median`}{The median result of the tuning.}
 \item{`optimal`}{The optimal configuration based on the minimum mean or median FDP over repetitions.
 Each component is a dataframe containing the configuration, performance metrics (including FDP and TPP), run times and average performance metrics.}
 }
}
\description{
Tuning function based on semi-simulation on a series of parameters in PLSKO, including ncomp, threshold.q, threshold.abs and sparsity. The function will return the optimal configuration based on the minimum mean or median FDP over repetitions.
}
\examples{
data(cfRNA_placenta)
X <- cfRNA_placenta$counts
cfRNA_tune <- plsko_tuning(X, p_s = 10, n_ko = 10, q = 0.05, parallel = TRUE, ncore = NULL, ncomp = seq(3, 9, 2), threshold.q = 0, threshold.abs = c(0.5, 0.3, 0), sparsity = 1, simpls = F, seed = 1, fdp.measure = "median", early.stop = F)
plot(cfRNA_tune)

}
